Revert to VS 2015 instead of VS 2019
Trying to use the VS 2019 boost-ci library for PyTango led to
"unresolved symbol" linker errors, as PyTango is compiled with VS 2015.
While VS C++ 2015, 2016 and 2019 are "compatible", compatibility
requires compilation with a version at least as new as the newest
component used.
https://docs.microsoft.com/en-us/cpp/porting/binary-compat-2015-2017?view=vs-2019

Keeping PyTango and its dependencies on the oldest compiler provides
maximum compatibilty for application developers.  They could use
VS 2015, 2017 or 2019.